recognition/tasks/duetface/requirements.txt
dareblopy
torchjpeg
torchsummary
shapely